backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
SQLAlchemy
aiosqlite
greenlet
httpx
PyJWT
passlib[bcrypt]
google-auth
requests
gunicorn
asyncpg
